requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-decouple==3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Pillow==11.0.0
redis==5.0.1
django-redis==5.4.0
